requirements.txt
alembic==1.8.0
aniso8601==9.0.1
attrs==21.4.0
bleach==5.0.0
cached-property==1.5.2
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
deepl==1.11.0
detectlanguage==1.5.0
dnspython==2.2.1
eventlet==0.33.1
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
furl==2.1.3
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.16.0
orderedmultidict==1.0.1
packaging==21.3
passlib==1.7.4
phonenumbers==8.12.50
pip-autoremove==0.10.0
platformdirs==2.5.2
psutil==5.9.5
psycopg2==2.9.3
pyjsonq==1.0.2
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.1
requests==2.28.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
SQLAlchemy==1.4.37
text-unidecode==1.3
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.2.3
zeep==4.1.0
zipp==3.8.0
